change the data type of account number in Customer_Bank_Account.

+++ src/main/java/com/IpayUsers/IpayUsers/Domain/Customer_Bank_Account.java
@@ -13,7 +13,7 @@ import lombok.NoArgsConstructor;
 @Data
 public class Customer_Bank_Account {
     @Id
-    private int accountNumber;
+    private long accountNumber;
     private String status;
     private boolean isDefault;
